fix windows 64bit build on vs2015
Visual Studio 2015 will report as follows at the time of 64bit build.

/core/vcl/win/gdi/winlayout.cxx(3972): error C2397:
conversion from 'long' to 'FLOAT' requires a narrowing.

Reviewed-on: https://gerrit.libreoffice.org/23367
Tested-by: Jenkins <ci@libreoffice.org>

+++ vcl/win/gdi/winlayout.cxx
@@ -3967,7 +3967,9 @@ bool D2DWriteTextOutRenderer::DrawGlyphs(const Point & origin, uint16_t * pGid, 
         {
             std::copy(pGid, pGidEnd, glyphIndices);
             std::copy_n(pAdvances, nGlyphs, glyphAdvances);
-            std::transform(pOffsets, pOffsets + nGlyphs, glyphOffsets, [](Point &o) { return DWRITE_GLYPH_OFFSET{ o.X(), o.Y() }; });
+            std::transform(pOffsets, pOffsets + nGlyphs,
+                glyphOffsets,
+                [](Point &o) { return DWRITE_GLYPH_OFFSET{static_cast<FLOAT>(o.X()), static_cast<FLOAT>(o.Y()) }; });
             D2D1_POINT_2F baseline = { origin.X() - bounds.Left(), origin.Y() - bounds.Top() };
             DWRITE_GLYPH_RUN glyphs = {
                 mpFontFace,
